% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopySeg_sequenza.R
\name{CopySeg_sequenza}
\alias{CopySeg_sequenza}
\title{Copynumber estimation}
\usage{
CopySeg_sequenza(x, AD, file.name)
}
\arguments{
\item{x}{A \code{vcfR} object of the sequencing calls.
The sample names can be queried from \code{x}.}

\item{AD}{a \code{character} deoning \emph{ID} for depth of the reference allele.
This is often separately present in the VCF file. Default is \code{NULL}.}

\item{file.name}{an optional \code{character} to define output file name. Default is \emph{tumor.sample}.}
}
\value{
A transformed \code{dataframe} usable in \emph{CloneStrat} that represents data on all variants
in the .vcf file. It returns summaries
on the variants with the collumn \emph{CN.profile} depicting the estimated allelic segmentations.
}
\description{
Allelic segmentations are estimated for one sample at a time with unfiltered sequencing calls
using the package \code{sequenza}.
This function can handle only a combination of one tumor sample with
a matched normal sample.
}
\details{
The function writes a \emph{.txt} data in working directory with the name defined in
\code{file.name} used by \emph{sequenza}. The output file written can be used in conjunction with
post variants call sequence file. These can be merged and used for surther analysis
with \code{\link{cluster.doc}} or \code{\link{seqn.scale}}
}
