% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutect2.qc.R
\name{mutect2.qc}
\alias{mutect2.qc}
\title{Quality Control on Mutect2 output}
\usage{
mutect2.qc(WES, sample.name)
}
\arguments{
\item{WES}{A \code{dataframe} of the Mutect2 output}

\item{sample.name}{a \code{vector} of sample names or IDs}
}
\value{
A transformed \code{dataframe} usable in \emph{CloneStrat} that represents data on each variant of each sample in rows
}
\description{
A quality control (QC) and transformation on the WES output from the Mutect2 variant caller. This re-organizes the data in a way that is
friendlier for using in \emph{CloneStrat}
}
\examples{
\donttest{#res<-mutect2.qc(WES,sample.name)}
}
