% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{primary_2}
\alias{primary_2}
\title{Human neuroblastoma tumor sample}
\format{
An object of class \code{vcfR} of dimension 149873 x 8 x 3.
}
\usage{
primary_2
}
\description{
DNA sample collected from a primary tumor site (different than that of \emph{primary_1}) was sequenced.
This is a pre-processing vcfR file used for variant calling.
}
\seealso{
\code{\link{Neuroblastoma}} \code{\link{primary_1}} \code{\link{metastasis_1}} \code{\link{metastasis_2}}

\href{https://doi.org/10.1038/s41588-018-0131-y}{Karlsson \emph{et al}., 2018}
}
\keyword{datasets}
