% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.plot.R
\name{segment.plot}
\alias{segment.plot}
\title{Plot of allelic composition}
\usage{
segment.plot(data, base.copy)
}
\arguments{
\item{data}{A \code{match.maker} or \code{\link{AlleleComp}} derived object.}

\item{base.copy}{is the baseline  balanced copynumber present in the sample usually "1 + 1" or "2 + 2".}
}
\value{
A plot of the allelic segmentation with average log-transformed coverage ratios in X-axis and
average allelic-imbalances in the Y-axis. This plot can be interpreted in the similar fashion as described
by \href{https://genomebiology.biomedcentral.com/articles/10.1186/gb-2011-12-10-r108}{Rasmussen \emph{et al}., 2011}
}
\description{
Departure in clusters of different allelic composition are portrayed for tumor sample
}
\examples{
\donttest{#segment.plot(data = data, base.copy = "1 + 1")}

}
\seealso{
\code{\link{match.maker}}, \code{\link{AlleleComp}}
}
