% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqn.scale.R
\name{seqn.scale}
\alias{seqn.scale}
\title{Probabilistic quotient normalization of DNA sequencing data}
\usage{
seqn.scale(x, vaf, CCF)
}
\arguments{
\item{x}{A \code{dataframe} containing summary from DNA sequencing with first column as sample IDs of corresponding variants.}

\item{vaf}{The column number of \code{x} that includes VAFs.}

\item{CCF}{The column number of \code{x} that includes CCFs.}
}
\value{
A \code{dataframe} with all the elements of \code{x} with the new estimated VAFs in the column
\emph{scaled.vaf} and an additional column \emph{unscaled.vaf} that includes the original VAFs.
}
\description{
A normalization technique based on cancer / tumor cell fractions of the samples sequenced to infer homogeneity
}
\details{
Probabilistic quotient normalization normalization technique described in
\emph{Dieterle, et al. (2006)} applied on the cancer cell fraction (CCF)
of respective samples to rescale variant allele frequencies (VAF) accordingly. The general idea is to put most
confidence in the sample with highest CCF and adjust the VAFs of other samples based on the departure in CCF
of the other samples from that with the highest.

This method is particularly suggested if the CCFs accross samples vary more than 10%.
}
\examples{
\donttest{#pqn.dat<-seqn.scale(test.dat,vaf=2,CCF=3)
#hist(pqn.dat$scaled.vaf)}
}
\seealso{
\code{\link{cluster.doc}}
}
