% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T.goodness.test.R
\name{T.goodness.test}
\alias{T.goodness.test}
\title{Test of fit of clonal deconvolution}
\usage{
T.goodness.test(x)
}
\arguments{
\item{x}{A \code{dataframe} with the first three columns in the specific order: sample name or ID of a variant, variant allele frquencies (VAF)
and cancer cell fraction (CCF)}
}
\value{
A list of two objects. \emph{x} is same as the input \code{dataframe} with addede columns named \emph{expected VAF_}, \emph{chi_sq_}
and \emph{P value_} corresponding to each cloud of clone : Sub-clone combination. \emph{rej} is a subset of \code{x} containing variants
that fail the test for at least one cloud.

\emph{expected VAF_} represents estimated variant allele frequencies for a given cloud.

\emph{chi_sq_} is the Chi square test statistic for the cloud.

\emph{P value_} is the P value corresponding to the \emph{chi_sq_} statistic.
}
\description{
A chi square test to assess the \emph{goodness of fit} of the clonal : sub-clonal clouds.
This test can be used to obtain outliers that do not fit into the proposed clonal deconvolution space.
}
\examples{
\donttest{
#test<-T.goodness.test(test.dat)
#head(test)}
}
