% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{test.dat}
\alias{test.dat}
\title{Random number generated WES data for eight hypothetical samples}
\format{
An object of class \code{"dataframe"}
}
\usage{
data(test.dat)
}
\value{
\code{sample} is column of IDs corresponding to 8 distinct samples.

\code{vaf} denotes the variant allele frequencies of each variant (see \code{annotation}).

\code{CCF} are the cancer cell fractions of each sample.

\code{annotation} indicates corresponding variants for which observations are notes in each row.
Variants can be shared among several samples as well as be private mutation.
}
\description{
Data generated with varying random normal probabilities.
Ideal llelic composition is assumed resulting in two separate distinct clouds of clones and sub-clones.
}
\examples{
\donttest{data(test.dat)
table(test.dat$CCF)
table(test.dat$annotation)
hist(test.dat$vaf)}
}
\keyword{datasets}
