% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.plot.R
\name{variant.plot}
\alias{variant.plot}
\title{Multi-sample variant plot}
\usage{
variant.plot(CD.obj, annotation.col, variant)
}
\arguments{
\item{CD.obj}{A \code{cluster.doc object}}

\item{annotation.col}{name of the column containing annotations of the variants
in original WES \code{dataframe} used in the clonal deconvolution using \code{cluster.doc}}

\item{variant}{a \code{character} string specifying \emph{only one} annotation which is to be displayed}
}
\value{
A plot object with the relevant annotation highlighted
}
\description{
Plotting a specific variant present in more than one WES sample
}
\examples{
\donttest{#cd.res<-cluster.doc(test.dat,1,2)
#variant.plot(cd.res,'annotation','variant_74')}
}
