% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View_summary.R
\name{View_summary}
\alias{View_summary}
\title{Copynumber plots}
\usage{
View_summary(data, filename = NULL)
}
\arguments{
\item{data}{An \code{\link{AlleleComp}} object of the sequencing calls.}

\item{filename}{a \code{character} denominating sample name.}
}
\value{
A pdf file with plots on twenty two autosomes. The left panel of each plot shows allelic imbalance against average segmental log ratios.
In the right panel from top to bottom three plots describes chromosomal relative coverage, allele frequency and estimated copynumber.
}
\description{
Summaries of copynumber estimations and allelic segmentations described in four plots
}
\examples{
\donttest{#met1<-AlleleComp(data = metastasis_1, AD = "AD", method = "naive", file = "TEMP");
#View_summary(data = met1, filename = "NB_met1")}

}
